--- app/controllers/books_controller.rb
@@ -1,5 +1,0 @@
-class BooksController < ApplicationController
-  def index
-    @books = Book.all
-  end
-end
