+++ app/javascript/components/Header.vue
@@ -1,0 +1,15 @@
+<template>
+    <div>
+        <nav>
+            <div class="nav-wrapper">
+                <router-link to="/" class="brand-logo">Bookshelf</router-link>
+                <ul id="nav-mobile" class="right">
+                    <li><router-link to="">本の登録</router-link></li>
+                </ul>
+            </div>
+        </nav>
+    </div>
+</template>
+
+<script></script>
+<style scoped></style>
